add milvus_cli to test cli
